1. Fixed a small bug in the spatial kdtree. 2. I've decided that printing the yt performance counters to the yt log is ok. So it does that now. 3. The padding for subvolumes in chainHOP is now done with MPI, rather than reading in the extra data off disk. This should substantially speed up unigrid data read times, and also help AMR datasets to a lesser degree.
--HG--
branch : hopwrap

+++ yt/extensions/spatial/ckdtree.pyx
@@ -591,7 +591,7 @@ cdef class cKDTree:
             If x has shape tuple+(self.m,), then i has shape tuple+(k,).
             Missing neighbors are indicated with self.n.
         """
-        cdef np.ndarray[int, ndim=2] ii
+        cdef np.ndarray[long, ndim=2] ii
         cdef np.ndarray[double, ndim=2] dd
         cdef np.ndarray[double, ndim=2] xx
         cdef np.ndarray[double, ndim=1] cperiod
